cleanup backbone.sync db object

+++ frontEnd/db.js
@@ -34,151 +34,168 @@ var Transcription = new LinvoDB(transcriptionModel, schema, options);
 
 var DB = {};
 
+// Setting up media folders for media and tmp media on local file system,
+// user libary application support folder
+var tmpMediaFolder = dataPath + '/tmp_media';
+var mediaFolder = dataPath + '/media';
+
+// if media folder does not exists create it
+if (!fs.existsSync(tmpMediaFolder)) {
+  console.debug('tmpMediaFolder folder not present, creating tmpMediaFolder folder');
+  fs.mkdirSync(tmpMediaFolder);
+} else {
+  // do nothing, build folder was already there
+  console.debug('tmpMediaFolder folder was already present');
+}
+
+// if temp media folder does not exists create it
+if (!fs.existsSync(mediaFolder)) {
+  console.debug('mediaFolder folder not present, creating mediaFolder folder');
+  fs.mkdirSync(mediaFolder);
+} else {
+  // do nothing, build folder was already there
+  console.debug('mediaFolder folder was already present');
+}
+
 /**
-* @function
-* @description Create functionality, mapped to REST POST
-* @param {object} method - Backbone RESTfull method request.
-* @param {object} model - Backbone model being handled.
-* @param {object} options - Sucess or failure callback.
-* @returns {object} sucess callback with backbone model containing db id
-*/
+ * Create a callback function for LinvoDB queries
+ * @param {function} success - Success callback
+ * @param {function} error - Error callback
+ * @returns {fuction} callback for LinvoDB
+ */
+function makeLinvoCallback(success, error) {
+  return function(err, found) {
+    if (err) {
+      // if we have an error, log it and bail
+      console.error(err);
+      return error(err);
+    }
+
+    // return transcription collection
+    success(found);
+  };
+}
+
+/**
+ * @function
+ * @description Create functionality, mapped to REST POST
+ * @param {object} method - Backbone RESTfull method request.
+ * @param {object} model - Backbone model being handled.
+ * @param {object} options - Sucess or failure callback.
+ * @returns {object} sucess callback with backbone model containing db id
+ */
 DB.create = function(model, success, error) {
+  console.debug('DB.create', model.constructor.modelType);
   if (model.constructor.modelType == 'transcription') {
     var newElement = model.toJSON();
     // create transcription element to save in db
     var transcription = new Transcription(newElement);
     // save transcription in db
     transcription.save(function(err) {
+      if (err) {
+        // if we have an error, log it and bail
+        console.error(err);
+        return error(err);
+      }
+
       // updating backbone with saved transcritpion, containing db id
       model.set(transcription);
+
       // returning saved transcription callback
       success(model);
-    });
 
-    // setting up media folders for media and tmp media on local file system,
-    // user libary application support folder
-    var tmpMediaFolder = dataPath + '/tmp_media';
-    var mediaFolder = dataPath + '/media';
-    // if media folder does not exists create it
-    if (!fs.existsSync(tmpMediaFolder)) {
-      console.log('tmpMediaFolder folder not present, creating tmpMediaFolder folder');
-      fs.mkdirSync(tmpMediaFolder);
-    } else {
-      // do nothing, build folder was already there
-      console.log('tmpMediaFolder folder was already present');
-    }
+      // using interactive_transcription_generator to generate metadata,
+      // transcription json
+      // webm video preview
+      transcription_generate({
+        id: transcription._id,
+        videoUrl: newElement.videoUrl,
+        title: newElement.title,
+        description: newElement.description,
+        // TODO: this is hardcoded, and this variable is not used, fix me!
+        // tmpWorkFolder:"/",
+        // destFolder:"/media",
+        tmpWorkFolder: tmpMediaFolder,
+        destFolder: mediaFolder,
+        keys: window.IBMWatsonKeys,
+        languageModel: newElement.languageModel,
+        sttEngine: newElement.sttEngine,
+        cbMetadata: function(resp) {
+          // update current transcription with metadata data
+          Transcription.findOne({ _id: resp.id }, function (err, trs) {
+            console.debug('got metadata for transcription: ' + trs._id);
+            trs.metadata = resp;
+            // saving current transcription
+            trs.save(function(err) { if (err) console.error(err); });
+          });
+        },
+        cbTranscription: function(resp) {
+          // updating current transcription with transcription json.
+          Transcription.findOne({ _id: resp.id }, function (err, trs) {
+            console.debug('got transcription json for transcription: ' + trs._id);
+            // updating transcription attributes with result
+            trs.audioFile = resp.audioFile;
+            trs.processedAudio = resp.processedAudio;
+            trs.text = resp.text;
+            trs.status = resp.status;
+            // saving current transcription
+            trs.save(function(err) { if (err) console.error(err); });
+          });
+        },
+        cbVideo: function(resp) {
+          // updating current transcription with webm html5 video preview.
+          Transcription.findOne({ _id: resp.id }, function (err, trs) {
+            console.debug('got video webm back for transcription: ' + trs._id);
+            // updating transcription attributes with result
+            trs.videoOgg = resp.videoOgg;
+            trs.processedVideo = resp.processedVideo;
+            // saving current transcription
+            trs.save(function(err) { if (err) console.error(err); });
+          });
+        }
+      });
 
-    // if temp media folder does not exists create it
-    if (!fs.existsSync(mediaFolder)) {
-      console.log('mediaFolder folder not present, creating mediaFolder folder');
-      fs.mkdirSync(mediaFolder);
-    } else {
-      // do nothing, build folder was already there
-      console.log('mediaFolder folder was already present');
-    }
-
-    // using interactive_transcription_generator to generate metadata,
-    // transcription json
-    // webm video preview
-    transcription_generate({
-      id: transcription._id,
-      videoUrl: newElement.videoUrl,
-      title: newElement.title,
-      description: newElement.description,
-      // TODO: this is hardcoded, and this variable is not used, fix me!
-      // tmpWorkFolder:"/",
-      // destFolder:"/media",
-      tmpWorkFolder: tmpMediaFolder,
-      destFolder: mediaFolder,
-      keys: window.IBMWatsonKeys,
-      languageModel: newElement.languageModel,
-      sttEngine: newElement.sttEngine,
-      cbMetadata: function(respM) {
-        // update current transcription with metadata data
-        Transcription.findOne({ _id: respM.id }, function (err, trs) {
-          console.info('got metadata for transcription: ' + transcription._id);
-          trs.metadata = respM;
-          // saving current transcription
-          trs.save(function(err) {
-            // app.trigger('updateTranscription:'+trs._id);
-          });
-        });
-      },
-      cbTranscription: function(resp) {
-        // updating current transcription with transcription json.
-        Transcription.findOne({ _id: resp.id }, function (err, trs) {
-          console.info('got transcription json for transcription: ' + trs._id);
-          // updating transcription attributes with result
-          trs.audioFile = resp.audioFile;
-          trs.processedAudio = resp.processedAudio;
-          trs.text = resp.text;
-          trs.status = resp.status;
-          // saving current transcription
-          trs.save(function(err) {
-            // app.trigger('updateTranscription:'+trs._id);
-          });
-        });
-      },
-      cbVideo: function(resp) {
-        // updating current transcription with webm html5 video preview.
-        Transcription.findOne({ _id: resp.id }, function (err, trs) {
-          console.info('got video webm back for transcription: ' + trs._id);
-          // updating transcription attributes with result
-          trs.videoOgg = resp.videoOgg;
-          trs.processedVideo = resp.processedVideo;
-          // saving current transcription
-          trs.save(function(err) {
-            // app.trigger('updateTranscription:' + trs._id);
-          });
-        });
-      }
     });
   }
 };
 
 /**
-* @function
-* @description Read functionality,Find all  and fine One, mapped to REST GET
-* @param {object} method - Backbone RESTfull method request.
-* @param {object} model - Backbone model being handled.
-* @param {object} options - Sucess or failure callback.
-* @returns {object} sucess callback with backbone model
-*/
+ * @function
+ * @description Read functionality,Find all  and fine One, mapped to REST GET
+ * @param {object} method - Backbone RESTfull method request.
+ * @param {object} model - Backbone model being handled.
+ * @param {object} options - Sucess or failure callback.
+ * @returns {object} sucess callback with backbone model
+ */
 DB.read = function(model, success, error) {
   // If a colleciton
   if (model.models) {
-    console.info('Collection:' + model.constructor.modelType);
+    console.debug('DB.read', 'Collection', model.constructor.modelType);
     // for transcription model
     if (model.constructor.modelType == 'transcriptions') {
       // look in database
-      Transcription.find({}, function (err, transcriptions) {
-        // return transcription collection
-        success(transcriptions);
-      });
+      Transcription.find({}, makeLinvoCallback(success, error));
     }// if transcription collection
   } else {
-    // if a model
-    console.info('Model: ' + model.constructor.modelType);
+    console.debug('DB.read', 'Model', model.constructor.modelType, model._id);
     // for transcription model
     if (model.constructor.modelType == 'transcription') {
       // looks in database using transcription id
-      Transcription.findOne({ _id: model._id }, function (err, transcription) {
-          // return transcription model
-        success(transcription);
-      });
+      Transcription.findOne({ _id: model._id }, makeLinvoCallback(success, error));
     }
   }
 };
 
 /**
-* @function
-* @description Update functionality, mapped to REST PUT
-* @param {object} method - Backbone RESTfull method request.
-* @param {object} model - Backbone model being handled.
-* @param {object} options - Sucess or failure callback.
-* @returns {object} sucess callback with backbone model
-*/
+ * @function
+ * @description Update functionality, mapped to REST PUT
+ * @param {object} method - Backbone RESTfull method request.
+ * @param {object} model - Backbone model being handled.
+ * @param {object} options - Sucess or failure callback.
+ * @returns {object} sucess callback with backbone model
+ */
 DB.update = function(model, success, error) {
+  console.debug('DB.update', model.constructor.modelType);
   // for transcription model
   if (model.constructor.modelType == 'transcription') {
     // looks in database using transcription id
@@ -199,23 +216,21 @@ DB.update = function(model, success, error) {
       doc.audioFile           = model.attributes.audioFile;
       doc.metadata            = model.attributes.metadata;
       // saving transcription to database
-      doc.save(function(err) {
-        // returning sucess callback with backbone model to client side
-        success(model);
-      });
+      doc.save(makeLinvoCallback(success, error));
     });
   }
 };
 
 /**
-* @function
-* @description Patch functionality, mapped to REST PUT
-* @param {object} method - Backbone RESTfull method request.
-* @param {object} model - Backbone model being handled.
-* @param {object} options - Sucess or failure callback.
-* @returns {object} sucess callback with backbone model
-*/
+ * @function
+ * @description Patch functionality, mapped to REST PUT
+ * @param {object} method - Backbone RESTfull method request.
+ * @param {object} model - Backbone model being handled.
+ * @param {object} options - Sucess or failure callback.
+ * @returns {object} sucess callback with backbone model
+ */
 DB.patch = function(model, success, error) {
+  console.debug('DB.patch', model.constructor.modelType);
   if (model.constructor.modelType == 'transcription') {
     Transcription.findOne({ _id: model.get('_id') }, function(err, doc) {
       // TODO: there's got to be a better way to do this
@@ -232,60 +247,59 @@ DB.patch = function(model, success, error) {
       doc.audioFile           = model.attributes.audioFile;
       doc.metadata            = model.attributes.metadata;
 
-      doc.save(function(err) {
-        success(model);
-      });
+      doc.save(makeLinvoCallback(success, error));
     });
   }
 };
 
 /**
-* @function
-* @description Delete functionality, mapped to REST Delete
-* @param {object} method - Backbone RESTfull method request.
-* @param {object} model - Backbone model being handled.
-* @param {object} options - Sucess or failure callback.
-* @returns {object} sucess callback with backbone model
-*/
+ * @function
+ * @description Delete functionality, mapped to REST Delete
+ * @param {object} method - Backbone RESTfull method request.
+ * @param {object} model - Backbone model being handled.
+ * @param {object} options - Sucess or failure callback.
+ * @returns {object} sucess callback with backbone model
+ */
 DB.delete = function(model, success, error) {
+  console.debug('DB.delete', model.constructor.modelType);
   // /for transcription model
   if (model.constructor.modelType == 'transcription') {
-    console.log('autoEdit2API');
-    console.log(model.attributes.title);
-    console.log(model.attributes._id);
     // looks in database using transcription id
     // worth looking into alternative
     // https://github.com/Ivshti/linvodb3#removing-from-the-collection
     Transcription.remove({ _id: model._id }, {multi: false }, function (err, numRemoved) {
-      // removing media associated with transcription.
-      // TODO: this only deletes the video if the video has done processing. think about refactoring so that attribute can be present before processing starts. As it is now this means incomplete vidoes are left in the folder if the transcription is deleted
-      if (model.attributes.processedVideo) {
-          // TODO: review if this is the right way to check if the file exists before deleting it.
-        if (fs.existsSync(model.attributes.videoOgg)) {
+      if (err) {
+        console.error(err);
+        error(err);
+      } else if (numRemoved < 1) {
+        var msg = "Couldn't delete transcription " + model._id;
+        console.error(msg);
+        error(msg);
+      } else {
+        // removing media associated with transcription.
+        // TODO: this only deletes the video if the video has done processing. think about refactoring so that attribute can be present before processing starts. As it is now this means incomplete vidoes are left in the folder if the transcription is deleted
+        if (model.attributes.processedVideo && fs.existsSync(model.attributes.videoOgg)) {
           fs.unlinkSync(model.attributes.videoOgg);
         }
-      }
         // if the trascription has been shown the audio will always be present because is needed to generate the text transcription. But having test here just in case
-      if (model.attributes.audioFile) {
-        if (fs.existsSync(model.attributes.audioFile)) {
+        if (model.attributes.audioFile && fs.existsSync(model.attributes.audioFile)) {
           fs.unlinkSync(model.attributes.audioFile);
         }
-      }
         // returns sucess callback
-      console.log(model.attributes._id);
-      success(model);
-      // });
+        console.debug('Deleted transcription ' + model.attributes._id);
+        success(model);
+      }
     });
   }
 };
 
 /**
-* @function
-* @description API Object to override Backnone.sync
-* @param {object} method - Backbone RESTfull method request.
-* @param {object} model - Backbone model being handled.
-* @param {object} options - Sucess or failure callback.
-*/
+ * @function
+ * @description API Object to override Backnone.sync
+ * @param {object} method - Backbone RESTfull method request.
+ * @param {object} model - Backbone model being handled.
+ * @param {object} options - Sucess or failure callback.
+ */
 module.exports = function(method, model, options) {
   DB[method](model, options.success, options.error);
 };
